refactor plan selection into focused submodules (#22)
